calculate seasonal bd metrics

This function calculates various biodiversity metrics (richness and diversity) for a given tiny forest and class

calc_bd_metrics(df, class)

Arguments

class

The class for which the metrics need to be calculated

tfid

The tiny forest id for which the metrics need to be calculated

Value

A list containing the following:

metrics : data.frame containing monthly values of the metrics matrix : data.frame containing monthly count of species for each site, in wide format plot : a ggplot object showing monthly variation in richness

Examples

calc_bd_metrics(85, 'Aves')
#> Loading required package: permute
#> Loading required package: lattice
#> This is vegan 2.6-4
#> ── Attaching core tidyverse packages ──────────────────────── tidyverse 2.0.0 ──
#> ✔ dplyr     1.1.4     ✔ readr     2.1.5
#> ✔ forcats   1.0.0     ✔ stringr   1.5.1
#> ✔ ggplot2   3.5.1     ✔ tibble    3.2.1
#> ✔ lubridate 1.9.3     ✔ tidyr     1.3.1
#> ✔ purrr     1.0.2     
#> ── Conflicts ────────────────────────────────────────── tidyverse_conflicts() ──
#> ✖ dplyr::between()     masks data.table::between()
#> ✖ dplyr::filter()      masks stats::filter()
#> ✖ dplyr::first()       masks data.table::first()
#> ✖ lubridate::hour()    masks data.table::hour()
#> ✖ lubridate::isoweek() masks data.table::isoweek()
#> ✖ dplyr::lag()         masks stats::lag()
#> ✖ dplyr::last()        masks data.table::last()
#> ✖ lubridate::mday()    masks data.table::mday()
#> ✖ lubridate::minute()  masks data.table::minute()
#> ✖ lubridate::month()   masks data.table::month()
#> ✖ lubridate::quarter() masks data.table::quarter()
#> ✖ lubridate::second()  masks data.table::second()
#> ✖ purrr::transpose()   masks data.table::transpose()
#> ✖ lubridate::wday()    masks data.table::wday()
#> ✖ lubridate::week()    masks data.table::week()
#> ✖ lubridate::yday()    masks data.table::yday()
#> ✖ lubridate::year()    masks data.table::year()
#> ℹ Use the conflicted package to force all conflicts to become errors
#> Error: Argument 'x' to 'setDT' should be a 'list', 'data.frame' or 'data.table'