Calculate normalized difference vegetation index (NDVI) within a buffer around a point location

This function takes in a latitude, longitude, buffer distance, start and end time period, and cloud cover percentage, and returns a map of NDVI within the buffer region, along with statistics on the NDVI values within the buffer.

Usage

calc_ndvi_buff(
  lat = 25.1972,
  lon = 55.2744,
  dist = 500,
  start_date = "2021-01-01",
  end_date = "2021-12-31",
  cloud_cover = 10
)

Arguments

lat

Numeric. The latitude of the point location of interest.

lon

Numeric. The longitude of the point location of interest.

dist

Numeric. The distance in meters to draw a buffer around the point location.

start_date

Character. The start date in the format 'YYYY-MM-DD' for the time period of interest.

end_date

Character. The end date in the format 'YYYY-MM-DD' for the time period of interest.

cloud_cover

Numeric. The maximum allowable percentage of cloud cover in the satellite imagery used to calculate the NDVI. Default is 10.

Value

A list object containing the following components: map: A map with two layers displaying the NDVI and the input point location. tidy_dates: A tidied version of the image collection with start and end dates associated with each image in the collection#' image_ids: A vector of image IDs corresponding to the images used in the calculation. raster: A raster layer object of the NDVI values within the buffer region. ndvi_stats: Summary statistics of the NDVI values within the buffer region, including median, minimum, maximum, mean and standard deviation.

References

COPERNICUS/S2_SR_HARMONIZED image collection.

rgee package documentation: https://csaybar.github.io/rgee/index.html

tidyrgee package documentation: https://csaybar.github.io/tidyrgee/index.html

zoo package documentation: https://cran.r-project.org/web/packages/zoo/index.html

terra package documentation: https://rspatial.org/terr'

Examples

calc_ndvi_buff(lat = 25.1972, lon = 55.2744, dist = 500, start_date = '2021-01-01',
end_date = '2021-12-31',cloud_cover = 10)
#> Loading required package: reticulate
#> Loading required package: rgee
#> Loading required package: tidyrgee
#> 
#> Attaching package: ‘tidyrgee’
#> The following object is masked from ‘package:stats’:
#> 
#>     filter
#> The following object is masked from ‘package:graphics’:
#> 
#>     clip
#> Loading required package: zoo
#> 
#> Attaching package: ‘zoo’
#> The following objects are masked from ‘package:base’:
#> 
#>     as.Date, as.Date.numeric
#> Loading required package: terra
#> terra 1.7.29
#> 
#> Attaching package: ‘terra’
#> The following object is masked from ‘package:zoo’:
#> 
#>     time<-
#> The following object is masked from ‘package:tidyr’:
#> 
#>     extract
#> The following object is masked from ‘package:data.table’:
#> 
#>     shift
#> Error in py_discover_config(required_module, use_environment): Python specified in RETICULATE_PYTHON (Users/julianflowers/.virtualenvs/geemap/bin/python) does not exist