Retrieves occurrence data from NBN Atlas within a specified buffer around given longitude and latitude

Retrieves occurrence data from NBN Atlas within a specified buffer around given longitude and latitude

Usage

get_nbn_buffer(lon, lat, radius = 1, n = 10000)

Arguments

lon

numeric Longitude in decimal degrees

lat

numeric Latitude in decimal degrees

radius

numeric Buffer radius in kilometers, default is 1 km

n

numeric Maximum number of occurrences to retrieve, default is 10,000

Value

data frame containing occurrence information within the buffer

Examples

get_nbn_buffer(-1.5, 52.0, 2)
#> Loading required package: jsonlite
#> 
#> Attaching package: ‘jsonlite’
#> The following object is masked from ‘package:purrr’:
#> 
#>     flatten
#> Loading required package: tictoc
#> 
#> Attaching package: ‘tictoc’
#> The following objects are masked from ‘package:terra’:
#> 
#>     shift, size
#> The following object is masked from ‘package:data.table’:
#> 
#>     shift
#> 12.3 sec elapsed
#>     kingdom          phylum            classs         order            family
#> 1   Plantae    Tracheophyta     Magnoliopsida      Lamiales         Lamiaceae
#> 2  Animalia        Chordata              Aves Passeriformes      Fringillidae
#> 3   Plantae    Tracheophyta     Magnoliopsida       Apiales          Apiaceae
#> 4  Animalia      Arthropoda           Insecta    Coleoptera         Carabidae
#> 5  Animalia        Chordata              Aves Passeriformes          Turdidae
#> 6   Plantae    Tracheophyta     Magnoliopsida   Brassicales      Brassicaceae
#> 7   Plantae    Tracheophyta     Magnoliopsida        Poales           Poaceae
#> 8   Plantae    Tracheophyta     Magnoliopsida      Lamiales         Lamiaceae
#> 9  Animalia      Arthropoda           Insecta    Coleoptera         Carabidae
#> 10  Plantae    Tracheophyta     Magnoliopsida        Poales           Poaceae
#> 11 Animalia      Arthropoda           Insecta   Lepidoptera         Noctuidae
#> 12 Animalia        Chordata              Aves Passeriformes          Turdidae
#> 13  Plantae    Tracheophyta     Magnoliopsida      Lamiales         Lamiaceae
#> 14  Plantae    Tracheophyta     Magnoliopsida      Lamiales         Lamiaceae
#> 15    Fungi      Ascomycota     Leotiomycetes    Helotiales      Dermateaceae
#> 16 Animalia        Chordata              Aves Passeriformes     Troglodytidae
#> 17 Animalia        Chordata              Aves Passeriformes          Turdidae
#> 18  Plantae    Tracheophyta     Magnoliopsida    Sapindales       Sapindaceae
#> 19  Plantae    Tracheophyta     Magnoliopsida     Asterales        Asteraceae
#> 20 Animalia        Chordata              Aves Passeriformes      Aegithalidae
#> 21 Animalia        Chordata              Aves Columbiformes        Columbidae
#> 22 Animalia        Chordata              Aves    Piciformes           Picidae
#> 23  Plantae    Tracheophyta     Magnoliopsida     Asterales        Asteraceae
#> 24  Plantae    Tracheophyta     Magnoliopsida        Poales         Typhaceae
#> 25  Plantae    Tracheophyta     Magnoliopsida     Asterales        Asteraceae
#> 26  Plantae    Tracheophyta     Magnoliopsida  Ranunculales     Ranunculaceae
#> 27 Animalia        Chordata              Aves    Piciformes           Picidae
#> 28 Animalia        Chordata              Aves Passeriformes        Passeridae
#> 29  Plantae    Tracheophyta     Magnoliopsida      Lamiales         Lamiaceae
#> 30  Plantae    Tracheophyta     Magnoliopsida        Poales           Poaceae
#> 31  Plantae Marchantiophyta Jungermanniopsida  Metzgeriales     Metzgeriaceae
#> 32 Animalia        Chordata              Aves Passeriformes           Paridae
#> 33 Animalia        Chordata              Aves Passeriformes         Sturnidae
#> 34  Plantae    Tracheophyta     Magnoliopsida      Ericales       Primulaceae
#> 35 Animalia        Chordata              Aves Passeriformes      Fringillidae
#> 36    Fungi      Ascomycota   Lecanoromycetes    Lecideales       Lecideaceae
#> 37 Animalia        Chordata          Mammalia    Chiroptera  Vespertilionidae
#> 38 Animalia        Chordata              Aves Passeriformes        Passeridae
#> 39  Plantae    Tracheophyta     Magnoliopsida      Lamiales    Plantaginaceae
#> 40 Animalia        Chordata              Aves Passeriformes          Corvidae
#> 41 Animalia        Chordata              Aves Passeriformes     Troglodytidae
#> 42    Fungi      Ascomycota     Leotiomycetes    Helotiales Chlorociboriaceae
#> 43  Plantae    Tracheophyta     Magnoliopsida  Malpighiales     Euphorbiaceae
#> 44 Animalia      Arthropoda           Insecta   Trichoptera     Limnephilidae
#> 45 Animalia        Mollusca        Gastropoda     Pulmonata         Limacidae
#> 46  Plantae    Tracheophyta     Magnoliopsida      Lamiales         Lamiaceae
#> 47  Plantae    Tracheophyta     Magnoliopsida      Lamiales         Lamiaceae
#> 48  Plantae    Tracheophyta     Magnoliopsida        Poales           Poaceae
#> 49  Plantae    Tracheophyta     Magnoliopsida   Alismatales      Alismataceae
#> 50 Animalia        Chordata              Aves Passeriformes          Turdidae
#> 51 Animalia        Chordata              Aves Columbiformes        Columbidae
#> 52  Plantae    Tracheophyta     Magnoliopsida        Poales           Poaceae
#> 53 Animalia        Chordata              Aves Passeriformes       Emberizidae
#> 54 Animalia        Chordata              Aves Passeriformes           Paridae
#> 55  Plantae    Tracheophyta     Magnoliopsida    Geraniales       Geraniaceae
#> 56  Plantae    Tracheophyta     Magnoliopsida  Malpighiales        Salicaceae
#> 57 Animalia        Chordata              Aves Passeriformes          Corvidae
#> 58 Animalia        Chordata              Aves Passeriformes         Regulidae
#> 59  Plantae    Tracheophyta     Magnoliopsida       Fabales          Fabaceae
#> 60  Plantae    Tracheophyta     Magnoliopsida       Fabales          Fabaceae
#> 61 Animalia        Chordata              Aves    Piciformes           Picidae
#> 62 Animalia        Chordata              Aves Columbiformes        Columbidae
#> 63 Animalia        Chordata              Aves Passeriformes     Troglodytidae
#> 64  Plantae    Tracheophyta     Magnoliopsida       Fabales          Fabaceae
#> 65 Animalia        Chordata              Aves Passeriformes         Regulidae
#> 66 Animalia        Chordata              Aves Passeriformes          Turdidae
#> 67 Animalia        Chordata              Aves Passeriformes     Troglodytidae
#> 68 Animalia        Chordata              Aves Passeriformes         Sylviidae
#> 69 Animalia        Chordata              Aves Columbiformes        Columbidae
#> 70  Plantae    Tracheophyta     Magnoliopsida   Gentianales         Rubiaceae
#> 71  Plantae    Tracheophyta     Magnoliopsida       Apiales          Apiaceae
#>            genus decimalLatitude decimalLongitude year month
#> 1         Thymus        51.99435        -1.491659 1969  <NA>
#> 2      Fringilla        51.99881        -1.484326 1999    04
#> 3      Heracleum        52.00905        -1.512254 2017    09
#> 4   Pterostichus        52.01233        -1.491456 1960    01
#> 5         Turdus        51.99881        -1.484326 2006    12
#> 6        Sinapis        52.01233        -1.491456 1970  <NA>
#> 7       Trisetum        52.01245        -1.520595 1968  <NA>
#> 8         Mentha        52.01233        -1.491456 1970  <NA>
#> 9      Bembidion        52.01233        -1.491456 1960  <NA>
#> 10        Holcus        51.99447        -1.520787 1968  <NA>
#> 11     Bryophila        52.01233        -1.491456 1952    01
#> 12        Turdus        51.99881        -1.484326 1996    03
#> 13      Glechoma        51.99447        -1.520787 1975  <NA>
#> 14     Galeopsis        51.99447        -1.520787 1975  <NA>
#> 15      Mollisia        52.01233        -1.491456 1974    06
#> 16   Troglodytes        51.99881        -1.484326 2004    02
#> 17        Turdus        51.98982        -1.484429 2008    07
#> 18      Aesculus        52.01233        -1.491456 1976  <NA>
#> 19       Cirsium        51.99566        -1.483634 1990    07
#> 20    Aegithalos        51.99881        -1.484326 2001    12
#> 21       Columba        51.98982        -1.484429 2008    02
#> 22   Dendrocopos        51.99881        -1.484326 2004    01
#> 23     Pilosella        51.99435        -1.491659 1969  <NA>
#> 24    Sparganium        52.01233        -1.491456 1950  <NA>
#> 25    Tragopogon        52.01233        -1.491456 1950  <NA>
#> 26       Ficaria        51.99447        -1.520787 1976  <NA>
#> 27   Dendrocopos        51.98982        -1.484429 2007    03
#> 28        Passer        51.99881        -1.484326 1997    03
#> 29       Ballota        52.00780        -1.484223 2016    05
#> 30        Bromus        52.01233        -1.491456 1950  <NA>
#> 31     Metzgeria        51.98966        -1.476931 2006    08
#> 32     Cyanistes        51.99447        -1.520787 2013    01
#> 33       Sturnus        52.01233        -1.491456 1988    04
#> 34    Lysimachia        52.01233        -1.491456 1987  <NA>
#> 35       Chloris        51.99881        -1.484326 2001    01
#> 36      Porpidia        52.01233        -1.491456 1960  <NA>
#> 37        Myotis        51.99881        -1.484326 2004    06
#> 38        Passer        51.98982        -1.484429 2007    04
#> 39      Veronica        51.99447        -1.520787 1976  <NA>
#> 40       Coloeus        51.98982        -1.484429 2008    07
#> 41   Troglodytes        51.99881        -1.484326 1998    08
#> 42 Chlorociboria        52.01233        -1.491456 1991    10
#> 43     Euphorbia        52.01233        -1.491456 1976  <NA>
#> 44   Limnephilus        52.00787        -1.498791 1988    09
#> 45     Lehmannia        52.01233        -1.491456 1965  <NA>
#> 46       Ballota        52.01233        -1.491456 1968  <NA>
#> 47        Salvia        52.01233        -1.491456 1970  <NA>
#> 48         Briza        52.01233        -1.491456 1970  <NA>
#> 49        Alisma        52.01233        -1.491456 1968  <NA>
#> 50        Turdus        51.98982        -1.484429 2007    03
#> 51  Streptopelia        51.99881        -1.484326 2000    09
#> 52      Koeleria        52.01233        -1.491456 1950  <NA>
#> 53      Emberiza        51.99881        -1.484326 2000    03
#> 54     Cyanistes        51.98982        -1.484429 2008    04
#> 55      Geranium        52.00806        -1.513576 2017    05
#> 56         Salix        52.01245        -1.520595 1979  <NA>
#> 57        Corvus        52.01233        -1.491456 1988    04
#> 58       Regulus        51.99881        -1.484326 2000    09
#> 59         Lotus        52.00909        -1.512203 2017    05
#> 60         Vicia        51.99435        -1.491659 1970  <NA>
#> 61   Dendrocopos        52.01233        -1.491456 1968    04
#> 62  Streptopelia        52.01233        -1.491456 1988    04
#> 63   Troglodytes        52.01233        -1.491456 1988    04
#> 64         Vicia        52.01233        -1.491456 1987  <NA>
#> 65       Regulus        52.01233        -1.491456 1981    11
#> 66        Turdus        51.99881        -1.484326 2002    08
#> 67   Troglodytes        52.01233        -1.491456 1988    04
#> 68        Sylvia        52.01233        -1.491456 1968    04
#> 69  Streptopelia        51.99881        -1.484326 2001    09
#> 70      Cruciata        52.01245        -1.520595 1979  <NA>
#> 71     Heracleum        52.01233        -1.491456 1969  <NA>
#>                                    dataProviderName
#> 1            Botanical Society of Britain & Ireland
#> 2                     British Trust for Ornithology
#> 3                  National Plant Monitoring Scheme
#> 4                         Biological Records Centre
#> 5                     British Trust for Ornithology
#> 6            Botanical Society of Britain & Ireland
#> 7            Botanical Society of Britain & Ireland
#> 8            Botanical Society of Britain & Ireland
#> 9                         Biological Records Centre
#> 10           Botanical Society of Britain & Ireland
#> 11                           Butterfly Conservation
#> 12                    British Trust for Ornithology
#> 13           Botanical Society of Britain & Ireland
#> 14           Botanical Society of Britain & Ireland
#> 15                      British Mycological Society
#> 16                    British Trust for Ornithology
#> 17                    British Trust for Ornithology
#> 18           Botanical Society of Britain & Ireland
#> 19       Thames Valley Environmental Records Centre
#> 20                    British Trust for Ornithology
#> 21                    British Trust for Ornithology
#> 22                    British Trust for Ornithology
#> 23           Botanical Society of Britain & Ireland
#> 24           Botanical Society of Britain & Ireland
#> 25           Botanical Society of Britain & Ireland
#> 26           Botanical Society of Britain & Ireland
#> 27                    British Trust for Ornithology
#> 28                    British Trust for Ornithology
#> 29           Botanical Society of Britain & Ireland
#> 30           Botanical Society of Britain & Ireland
#> 31                      British Bryological Society
#> 32                    British Trust for Ornithology
#> 33                    British Trust for Ornithology
#> 34           Botanical Society of Britain & Ireland
#> 35                    British Trust for Ornithology
#> 36                           British Lichen Society
#> 37                           Bat Conservation Trust
#> 38                    British Trust for Ornithology
#> 39           Botanical Society of Britain & Ireland
#> 40                    British Trust for Ornithology
#> 41                    British Trust for Ornithology
#> 42                      British Mycological Society
#> 43           Botanical Society of Britain & Ireland
#> 44                        Biological Records Centre
#> 45 Conchological Society of Great Britain & Ireland
#> 46           Botanical Society of Britain & Ireland
#> 47           Botanical Society of Britain & Ireland
#> 48           Botanical Society of Britain & Ireland
#> 49           Botanical Society of Britain & Ireland
#> 50                    British Trust for Ornithology
#> 51                    British Trust for Ornithology
#> 52           Botanical Society of Britain & Ireland
#> 53                    British Trust for Ornithology
#> 54                    British Trust for Ornithology
#> 55                 National Plant Monitoring Scheme
#> 56           Botanical Society of Britain & Ireland
#> 57                    British Trust for Ornithology
#> 58                    British Trust for Ornithology
#> 59                 National Plant Monitoring Scheme
#> 60           Botanical Society of Britain & Ireland
#> 61                    British Trust for Ornithology
#> 62                    British Trust for Ornithology
#> 63                    British Trust for Ornithology
#> 64           Botanical Society of Britain & Ireland
#> 65                    British Trust for Ornithology
#> 66                    British Trust for Ornithology
#> 67                    British Trust for Ornithology
#> 68                    British Trust for Ornithology
#> 69                    British Trust for Ornithology
#> 70           Botanical Society of Britain & Ireland
#> 71           Botanical Society of Britain & Ireland
#>                   speciesGroups              vernacularName
#> 1       Plants, FloweringPlants                        <NA>
#> 2                Animals, Birds                   Chaffinch
#> 3       Plants, FloweringPlants                     Hogweed
#> 4  Animals, Arthropods, Insects                        <NA>
#> 5                Animals, Birds                   Blackbird
#> 6       Plants, FloweringPlants                    Charlock
#> 7       Plants, FloweringPlants            Yellow Oat-grass
#> 8       Plants, FloweringPlants                  Spear Mint
#> 9  Animals, Arthropods, Insects                        <NA>
#> 10      Plants, FloweringPlants         Creeping Soft-grass
#> 11 Animals, Arthropods, Insects              Marbled Beauty
#> 12               Animals, Birds                 Song Thrush
#> 13      Plants, FloweringPlants                  Ground-ivy
#> 14      Plants, FloweringPlants          Common Hemp-nettle
#> 15                        Fungi                        <NA>
#> 16               Animals, Birds                        Wren
#> 17               Animals, Birds                   Blackbird
#> 18      Plants, FloweringPlants              Horse-chestnut
#> 19      Plants, FloweringPlants               Spear Thistle
#> 20               Animals, Birds             Long-tailed Tit
#> 21               Animals, Birds                  Woodpigeon
#> 22               Animals, Birds    Great Spotted Woodpecker
#> 23      Plants, FloweringPlants          Mouse-ear-hawkweed
#> 24      Plants, FloweringPlants         Unbranched Bur-reed
#> 25      Plants, FloweringPlants                Goat's-beard
#> 26      Plants, FloweringPlants            Lesser Celandine
#> 27               Animals, Birds    Great Spotted Woodpecker
#> 28               Animals, Birds               House Sparrow
#> 29      Plants, FloweringPlants             Black Horehound
#> 30      Plants, FloweringPlants                  Soft-brome
#> 31           Plants, Bryophytes          Whiskered Veilwort
#> 32               Animals, Birds                    Blue Tit
#> 33               Animals, Birds                    Starling
#> 34      Plants, FloweringPlants              Creeping-Jenny
#> 35               Animals, Birds                  Greenfinch
#> 36                        Fungi                        <NA>
#> 37             Animals, Mammals              Natterer's Bat
#> 38               Animals, Birds               House Sparrow
#> 39      Plants, FloweringPlants                   Brooklime
#> 40               Animals, Birds                     Jackdaw
#> 41               Animals, Birds                        Wren
#> 42                        Fungi                Green Elfcup
#> 43      Plants, FloweringPlants                  Sun Spurge
#> 44 Animals, Arthropods, Insects                        <NA>
#> 45            Animals, Molluscs                   Tree Slug
#> 46      Plants, FloweringPlants             Black Horehound
#> 47      Plants, FloweringPlants                  Wild Clary
#> 48      Plants, FloweringPlants               Quaking-grass
#> 49      Plants, FloweringPlants              Water-plantain
#> 50               Animals, Birds               Mistle Thrush
#> 51               Animals, Birds               Collared Dove
#> 52      Plants, FloweringPlants          Crested Hair-grass
#> 53               Animals, Birds                Yellowhammer
#> 54               Animals, Birds                    Blue Tit
#> 55      Plants, FloweringPlants Small-flowered Crane's-bill
#> 56      Plants, FloweringPlants                 Goat Willow
#> 57               Animals, Birds                        Rook
#> 58               Animals, Birds                   Goldcrest
#> 59      Plants, FloweringPlants         Bird's-Foot-Trefoil
#> 60      Plants, FloweringPlants                  Bush Vetch
#> 61               Animals, Birds    Great Spotted Woodpecker
#> 62               Animals, Birds               Collared Dove
#> 63               Animals, Birds                        Wren
#> 64      Plants, FloweringPlants                  Bush Vetch
#> 65               Animals, Birds                   Goldcrest
#> 66               Animals, Birds                 Song Thrush
#> 67               Animals, Birds                        Wren
#> 68               Animals, Birds              Garden Warbler
#> 69               Animals, Birds               Collared Dove
#> 70      Plants, FloweringPlants                   Crosswort
#> 71      Plants, FloweringPlants                     Hogweed
#>                        species
#> 1           Thymus polytrichus
#> 2            Fringilla coelebs
#> 3        Heracleum sphondylium
#> 4        Pterostichus diligens
#> 5                Turdus merula
#> 6             Sinapis arvensis
#> 7          Trisetum flavescens
#> 8               Mentha spicata
#> 9          Bembidion lunulatum
#> 10               Holcus mollis
#> 11         Bryophila domestica
#> 12           Turdus philomelos
#> 13          Glechoma hederacea
#> 14          Galeopsis tetrahit
#> 15                        <NA>
#> 16     Troglodytes troglodytes
#> 17               Turdus merula
#> 18      Aesculus hippocastanum
#> 19             Cirsium vulgare
#> 20         Aegithalos caudatus
#> 21            Columba palumbus
#> 22           Dendrocopos major
#> 23       Pilosella officinarum
#> 24          Sparganium emersum
#> 25        Tragopogon pratensis
#> 26               Ficaria verna
#> 27           Dendrocopos major
#> 28           Passer domesticus
#> 29               Ballota nigra
#> 30           Bromus hordeaceus
#> 31      Metzgeria consanguinea
#> 32         Cyanistes caeruleus
#> 33            Sturnus vulgaris
#> 34       Lysimachia nummularia
#> 35             Chloris chloris
#> 36        Porpidia tuberculosa
#> 37            Myotis nattereri
#> 38           Passer domesticus
#> 39         Veronica beccabunga
#> 40            Coloeus monedula
#> 41     Troglodytes troglodytes
#> 42 Chlorociboria aeruginascens
#> 43       Euphorbia helioscopia
#> 44         Limnephilus ignavus
#> 45         Lehmannia marginata
#> 46               Ballota nigra
#> 47            Salvia verbenaca
#> 48                 Briza media
#> 49    Alisma plantago-aquatica
#> 50           Turdus viscivorus
#> 51       Streptopelia decaocto
#> 52          Koeleria macrantha
#> 53         Emberiza citrinella
#> 54         Cyanistes caeruleus
#> 55           Geranium pusillum
#> 56                Salix caprea
#> 57           Corvus frugilegus
#> 58             Regulus regulus
#> 59                        <NA>
#> 60                Vicia sepium
#> 61           Dendrocopos major
#> 62       Streptopelia decaocto
#> 63     Troglodytes troglodytes
#> 64                Vicia sepium
#> 65             Regulus regulus
#> 66           Turdus philomelos
#> 67     Troglodytes troglodytes
#> 68                Sylvia borin
#> 69       Streptopelia decaocto
#> 70           Cruciata laevipes
#> 71       Heracleum sphondylium
#>  [ reached 'max' / getOption("max.print") -- omitted 9929 rows ]