Get S2 Landcover

Function to retrieve Sentinel-2 landcover data for a specified location and time period.

get_s2_landcover(
  tf_id = 85,
  lon = -0.25083,
  lat = 52.5739214,
  dist = 1000,
  start_date = "2023-01-01",
  end_date = "2023-09-01"
)

Arguments

tf_id

Numeric ID associated with the training forest default = tf_id.

lon

Numeric, longitude coordinate of the location default = -0.250830.

lat

Numeric, latitude coordinate of the location default = 52.5739214.

dist

Numeric, buffer distance in meters around the location default = 1000.

start_date

Character string, start date of the time period in 'YYYY-MM-DD' format default = '2023-01-01'.

end_date

Character string, end date of the time period in 'YYYY-MM-DD' format default = '2023-09-01'.

Examples

## Specify the tf_id
tf_id <- 85

## Specify the location coordinates
lon <- -0.250830
lat <- 52.5739214

## Specify the buffer distance
dist <- 1000

## Specify the time period
start_date <- '2023-01-01'
end_date <- '2023-09-01'

## Call the get_s2_landcover function
get_s2_landcover(tf_id = tf_id, lon = lon, lat = lat, dist = dist)
#> Loading required package: rgee
#> Loading required package: tidyrgee
#> 
#> Attaching package: ‘tidyrgee’
#> The following object is masked from ‘package:stats’:
#> 
#>     filter
#> The following object is masked from ‘package:graphics’:
#> 
#>     clip
#> Loading required package: stars
#> Loading required package: abind
#> Loading required package: sf
#> Linking to GEOS 3.11.0, GDAL 3.5.3, PROJ 9.1.0; sf_use_s2() is TRUE
#> Loading required package: terra
#> terra 1.7.71
#> 
#> Attaching package: ‘terra’
#> The following objects are masked from ‘package:tictoc’:
#> 
#>     shift, size
#> The following object is masked from ‘package:glue’:
#> 
#>     trim
#> The following object is masked from ‘package:janitor’:
#> 
#>     crosstab
#> The following object is masked from ‘package:tidyr’:
#> 
#>     extract
#> The following object is masked from ‘package:data.table’:
#> 
#>     shift
#> Error in py_discover_config(required_module, use_environment): Python specified in RETICULATE_PYTHON (Users/julianflowers/.virtualenvs/thesis-final/bin/python) does not exist