Plot OS Land Cover

Generate an interactive map of Ordnance Survey (OS) Land Cover data around a specified coordinate using the OS Data Hub API.

Usage

plot_os_land_cover(lon, lat, key, buff = 1000, n = 5000L)

Arguments

lon

numeric value indicating the longitude of the required point (e.g. -0.1426)

lat

numeric value indicating the latitude of the required point (e.g. 51.5244)

key

character value for the OS Data Hub API key in use

buff

numeric value indicating the buffer around the specified point in metres (default = 1000)

n

integer value indicating the maximum number of results to return (default = 10000)

Value

returns a list including the following objects:

sf: an sf object containing the queried data,

map: an interactive map generated using the queried data,

tiers: an sf object containing the tier B land cover information.

Details

Requires the OS Data Hub API key and needed packages to run (tidyverse, reticulate, mapview, sf, tictoc, colourspace, leaflet).

Examples

# Generate a plot of land cover around the London Eye
key <- "insert OS Data Hub API key"
plot_os_land_cover(lon = -0.1194, lat = 51.5033, key = key)
#> Error: requests.exceptions.HTTPError: 400 Client Error: Bad Request for url: https://api.os.uk/features/ngd/ofa/v1/collections/lnd-fts-land-1/items/?filter=INTERSECTS%28geometry%2C+POLYGON+%28%28999.8806+-948.4967%2C+999.8806+1051.5033%2C+-1000.1194+1051.5033%2C+-1000.1194+-948.4967%2C+999.8806+-948.4967%29%29%29&filter-crs=http%3A%2F%2Fwww.opengis.net%2Fdef%2Fcrs%2FEPSG%2F0%2F27700&limit=100&offset=0